+++ lib/middleman-google-analytics/extension.rb
@@ -1,67 +1,50 @@
+require 'uglifier'
+
 module Middleman
-  module GoogleAnalytics
-    class Options < Struct.new(:tracking_id, :anonymize_ip, :allow_linker, :domain_name, :debug); end
+  class GoogleAnalyticsExtension < Extension
+    option :tracking_id, nil, 'Property ID'
+    option :anonymize_ip, false, 'Removing the last octet of the IP address'
+    option :domain_name, nil, 'Tracking across a domain and its subdomains'
+    option :allow_linker, false, 'Tracking across multiple domains and ' \
+                                 'subdomains'
+    option :debug, false, 'Tracking Code Debugger'
+    option :development, true, 'Tracking in development environment'
+    option :minify, false, 'Compress the JavaScript code'
 
-    class << self
-      def options
-        @@options ||= Options.new
-      end
+    def initialize(app, options_hash={}, &block)
+      super
 
-      def registered(app, options={})
-        @@options ||= Options.new(*options.values_at(*Options.members))
-        yield @@options if block_given?
-
-        if @@options.allow_linker and not @@options.domain_name
-          $stderr.puts 'Google Analytics: Please specify a domain_name when using allow_linker'
-          raise 'No domain_name given'
-        end
-
-        app.send :include, InstanceMethods
-      end
-      alias :included :registered
+      app.set :google_analytics_settings, options
     end
 
-    module InstanceMethods
+    def after_configuration
+      unless options.tracking_id
+        $stderr.puts 'Google Analytics: Please specify a property ID'
+        raise 'No property ID given'
+      end
+
+      if options.allow_linker and not options.domain_name
+        $stderr.puts 'Google Analytics: Please specify a domain_name when ' \
+                     'using allow_linker'
+        raise 'No domain_name given'
+      end
+    end
+
+    helpers do
       def google_analytics_tag
-        options = ::Middleman::GoogleAnalytics.options
-        options.debug = development? if options.debug.nil?
-        ga = options.debug ? 'u/ga_debug' : 'ga'
-        domain_name = options.domain_name
-        if tracking_id = options.tracking_id
-          gaq = []
-          gaq << ['_setAccount', "#{tracking_id}"]
-          gaq << ['_setDomainName', "#{domain_name}"] if domain_name
-          gaq << ['_setAllowLinker', true] if options.allow_linker
-          gaq << ['_gat._anonymizeIp'] if options.anonymize_ip
-          gaq << ['_trackPageview']
-          %Q{<script type="text/javascript">
-  var _gaq = _gaq || [];
-  #{gaq.map! { |x| "_gaq.push(#{x});" }.join("\n  ")}
-  (function() {
-    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-    ga.src = ('https:' == document.location.protocol ? '//ssl' : '//www') + '.google-analytics.com/#{ga}.js';
-    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-  })();
-</script>}
-        end
+        @options = google_analytics_settings
+        file = File.join(File.dirname(__FILE__), 'ga.js.erb')
+        content = ERB.new(File.read(file)).result(binding)
+        content = Uglifier.compile(content) if google_analytics_settings.minify
+        content_tag(:script, content, type: 'text/javascript')
       end
 
       def google_analytics_universal_tag
-        options = ::Middleman::GoogleAnalytics.options
-        options.debug = development? if options.debug.nil?
-        ga = options.debug ? 'u/ga_debug' : 'ga'
-        domain_name = options.domain_name
-        if tracking_id = options.tracking_id
-          %Q{<script type="text/javascript">
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-  })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-  ga('create', '#{ tracking_id }', '#{ domain_name || 'auto' }'#{ options.allow_linker ? ", {'allowLinker': true}" : '' });#{
-    options.anonymize_ip ? "\n  ga('set', 'anonymizeIp', true);" : '' }
-  ga('send', 'pageview');
-</script>}
-        end
+        @options = google_analytics_settings
+        file = File.join(File.dirname(__FILE__), 'analytics.js.erb')
+        content = ERB.new(File.read(file)).result(binding)
+        content = Uglifier.compile(content) if google_analytics_settings.minify
+        content_tag(:script, content, type: 'text/javascript')
       end
     end
   end